% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{getRequestBody}
\alias{getRequestBody}
\title{Get request body from request object}
\usage{
getRequestBody(req)
}
\arguments{
\item{req}{The request object}
}
\value{
The content, correctly parsed according to Content-Type. If there
is no content, the function returns NULL
}
\description{
Get request body from request object
}
