% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{setToken}
\alias{setToken}
\title{Confirm the user's authentication status}
\usage{
setToken(cookie)
}
\arguments{
\item{cookie}{Character cookie from the request object}
}
\value{
Logical: \code{TRUE}/\code{FALSE} whether authed.
}
\description{
Given the request, determine whether the user is authenticated with
Crunch. Set the token if so.
}
